requirements.txt
discord.py
asyncpg
pytest
pytest-asyncio